+++ releases/dogwood/3/fun/config/lms/docker_run_production.py
@@ -1514,3 +1514,6 @@ PAYMENT_ADMIN = "paybox@fun-mooc.fr"
 # Force Edx to use `libcast_xblock` as default video player
 # in the studio (big green button) and if any xblock is called `video`
 XBLOCK_SELECT_FUNCTION = prefer_fun_video
+
+# AT Internet web analytics tracking
+AT_INTERNET_SMARTTAG = config("AT_INTERNET_SMARTTAG", default=None)